+++ src/ts/wysiwyg/processKeydown.ts
@@ -3,7 +3,7 @@ import {isCtrl, isFirefox} from "../util/compatibility";
 import {scrollCenter} from "../util/editorCommonEvent";
 import {
     fixBlockquote, fixCJKPosition,
-    fixCodeBlock, fixCursorDownInlineMath, fixDelete, fixHR,
+    fixCodeBlock, fixCursorDownInlineMath, fixDelete, fixFirefoxArrowUpTable, fixHR,
     fixList,
     fixMarkdown,
     fixTab,
@@ -301,12 +301,17 @@ export const processKeydown = (vditor: IVditor, event: KeyboardEvent) => {
     }
 
     if (isFirefox() && range.startOffset === 1 && startContainer.textContent.indexOf(Constants.ZWSP) > -1 &&
-        startContainer.previousSibling.nodeType !== 3 &&
+        startContainer.previousSibling && startContainer.previousSibling.nodeType !== 3 &&
         (startContainer.previousSibling as HTMLElement).tagName === "CODE" &&
         (event.key === "Backspace" || event.key === "ArrowLeft")) {
         // https://github.com/Vanessa219/vditor/issues/410
         range.selectNodeContents(startContainer.previousSibling);
         range.collapse(false);
+        event.preventDefault();
+        return true;
+    }
+
+    if (fixFirefoxArrowUpTable(event, blockElement, range)) {
         event.preventDefault();
         return true;
     }